Separated port map utilities into another file
