+++ src/GNUmakefile
@@ -56,6 +56,7 @@ OBJECTS += \
 	extension/kompat/kompat.o \
 	extension/fake_id0/fake_id0.o \
 	extension/portmap/portmap.o \
+	extension/portmap/map.o \
 	loader/loader-wrapped.o
 
 define define_from_arch.h
