+++ src/extension/extension.h
@@ -29,6 +29,7 @@
 
 #include "tracee/tracee.h"
 #include "syscall/seccomp.h"
+#include "extension/portmap/portmap.h"
 
 /* List of possible events.  */
 typedef enum {
@@ -181,9 +182,13 @@ extern int fake_id0_callback(Extension *extension, ExtensionEvent event, intptr_
 extern int care_callback(Extension *extension, ExtensionEvent event, intptr_t d1, intptr_t d2);
 
 /* Added extensions.  */
-extern int portmap_callback(Extension *extension, ExtensionEvent event, intptr_t d1, intptr_t d2);
+/**
+ * We use a global variable in order to support multiple port mapping options,
+ * otherwise we would have a different extension instance for each (port_in, port_out) pair,
+ * which would be a waste of memory and performance.
+ * This variable is modified only once, in the INITIALIZATION event.
+ */
 extern Extension *global_portmap_extension;
-int add_portmap_entry(uint16_t port_in, uint16_t port_out);
-int activate_netcoop_mode();
+extern int portmap_callback(Extension *extension, ExtensionEvent event, intptr_t d1, intptr_t d2);
 
 #endif /* EXTENSION_H */
